+++ app/screens/OnboardingProfileInfo/presenter.js
@@ -1,5 +1,4 @@
 import { connect as _connect } from 'react-redux'
-import { setUserNameInput } from '@state/actions/forms'
 import { saveUserName } from '@state/actions/profile'
 
 const mapStateToProps = (state) => ({})
